+++ .firebase/hosting.YnVpbGQ.cache
@@ -1,27 +1,26 @@
-sitemap.xml,1743866861611,0ff2cba8f04607521a4cff4b2f12529b139da1a5554f1c7e12a7fea84b815763
-site.webmanifest,1743819939359,65748f57e3a355ca3b5213dc1a8ab66a2a605f966aa3576f0c482129ce284b00
-robots.txt,1743866896448,36984a01bd543df4c81dd83fac1024d22f785c061d31473c29602025c55c67f2
-manifest.json,1743818994800,ee04fb47e525c67d8424ffe9b4d8a8a24e434504478afca4e0ca602146836d4c
-logo512.png,1743819865451,883b9820cf90019f2d84dc41c5a9045a50c947110026fa9907e9b6de4b468143
-logo192.png,1743819865451,883b9820cf90019f2d84dc41c5a9045a50c947110026fa9907e9b6de4b468143
-favicon.ico,1743819939312,a16989d4c5d5089f16b4789aa338d43d7c300a81ebc26c530d13830346ae0c9e
-favicon-32x32.png,1743819939249,cd66b345f0fe7eeecba70e882c99e69e92ded9a9333fe08061f0fec328263f73
-favicon-16x16.png,1743819939281,b7c55efedc5b8e611d5732c0ecbbd0df87e4a1cd161307eb634a567c68de0e07
-apple-touch-icon.png,1743819939062,6d6a75bece03595f34c90f74b45cf086588496aa8aeeae610ad55f294959fc8e
-android-chrome-512x512.png,1743819939546,9cc4ddd3ad0e6472d2d5548ef3fd4809fc5600dbbefa129d91f3b864ed4c1619
-android-chrome-192x192.png,1743819939390,ec29c5b4fa2e4d776499fc3bf63d7b9995b6d6a947b4683354492297d20fa94f
-algospectra-logo - Copy.jpeg,1743819865451,883b9820cf90019f2d84dc41c5a9045a50c947110026fa9907e9b6de4b468143
-404.html,1743829811304,af43c4b30327cccfd8f6344da70afc1cb2abbac4bb401a27a9bb15278d0e731b
-robots.txt,1743903820190,b2090cf9761ef60aa06e4fab97679bd43dfa5e5df073701ead5879d7c68f1ec5
-index.html,1743905094722,3325d30c5f85cc213a4a3e956d22024aef04953f1d31512c60887ee408372880
-asset-manifest.json,1743905094722,d83d810a9da8bc3ab3c3df53f1754cdbf6bb3436235ca4b776b0ea4dff45c9e3
-static/js/main.602a7709.js.LICENSE.txt,1743905094743,6a3d81f51aabfa2418776c74318e64f7e398fb657deaf532d0fde0ff7d587367
-static/js/453.8a995f6c.chunk.js.map,1743905094743,8696b11a30e32b1632e56dcbe74135f677691af51459e7308ca0ba6fc6925c07
-static/js/453.8a995f6c.chunk.js,1743905094743,0fa9890aff3cb336436289a125a670884917a1e9ff1bae4faa0bc503f90bedd5
-static/css/main.03748af8.css.map,1743905094745,23ea53fa20dcc6ce1db054412970856ea2fc8a338ccaa60ec78705c80e976b1d
-static/css/main.03748af8.css,1743905094743,d23cde5ac7a9db0535bf5c4c1eb2ea4482b74ef1421c5d6cfbc2862ac9115068
-static/media/god-dna.a807991fe1af48598719.jfif,1743905094743,0d07a05b971b8521f465e162853557910d3930e94ffe15ac7f7926a7b349750c
-static/js/main.602a7709.js,1743905094743,91461a8dd3e6fc1261ac215a210e728bf1ac470ee19f7af4338d27c4a8cdb2e4
-logo512.png,1743903820190,212b102aa09e51b3b3e06647e81f7801a61333e171f6582e8124379aabccb41d
-logo192.png,1743903820190,79e2b749561016bc8af300ea19f48347ceed3cb1a54f48ae456172eca45e08f0
-static/js/main.602a7709.js.map,1743905094746,729fe4287ceaf65cfac4b087be5c4f17a91a7c9cc6cd2e3f583bde54c71542d2
+sitemap.xml,1743977576018,0ff2cba8f04607521a4cff4b2f12529b139da1a5554f1c7e12a7fea84b815763
+site.webmanifest,1743977576018,65748f57e3a355ca3b5213dc1a8ab66a2a605f966aa3576f0c482129ce284b00
+robots.txt,1743977576018,27d58fba4b78b0651a1a7d19e163828e827b13742cb44ef61316d09b9f90902f
+manifest.json,1743977576018,4368aeaf848ae2e048765562c289452f33ad2a175c4b1951ea8bdf2ada0d5b10
+index.html,1743980848881,e77dd7eafdd8d619eedf35b9a0c8521e3f73050f4ae117dd6740f46a5c249407
+google75a7a43ef7ed76e0.html,1743977576003,8c0c0c33247076cb97876c056c90c0eb0dc54a4a660384236ba913a9b39489b0
+asset-manifest.json,1743980848881,910a7d0fc7a97b8f543e6b01451446cc0834220f11788f36bd081b7f5f67a59a
+favicon-32x32.png,1743977576003,cd66b345f0fe7eeecba70e882c99e69e92ded9a9333fe08061f0fec328263f73
+favicon-16x16.png,1743977576003,b7c55efedc5b8e611d5732c0ecbbd0df87e4a1cd161307eb634a567c68de0e07
+200.html,1743980859543,e77dd7eafdd8d619eedf35b9a0c8521e3f73050f4ae117dd6740f46a5c249407
+favicon.ico,1743977576003,a16989d4c5d5089f16b4789aa338d43d7c300a81ebc26c530d13830346ae0c9e
+static/js/main.25fbd85a.js.LICENSE.txt,1743980848896,6a3d81f51aabfa2418776c74318e64f7e398fb657deaf532d0fde0ff7d587367
+404.html,1743977576003,af43c4b30327cccfd8f6344da70afc1cb2abbac4bb401a27a9bb15278d0e731b
+static/js/453.8a995f6c.chunk.js.map,1743980848896,8696b11a30e32b1632e56dcbe74135f677691af51459e7308ca0ba6fc6925c07
+static/js/453.8a995f6c.chunk.js,1743980848896,0fa9890aff3cb336436289a125a670884917a1e9ff1bae4faa0bc503f90bedd5
+static/css/main.3775cc80.css,1743980848896,38548a22efc3d8869a2526a4cf9bd2414120227933d6f06f1eaf90836901f399
+static/css/main.3775cc80.css.map,1743980848896,5ac9b5600b6df6c71ce39379b284c6d798e07fc9cd7eca7df7ccc33d3ced1af9
+logo512.png,1743977576018,883b9820cf90019f2d84dc41c5a9045a50c947110026fa9907e9b6de4b468143
+android-chrome-192x192.png,1743977576003,ec29c5b4fa2e4d776499fc3bf63d7b9995b6d6a947b4683354492297d20fa94f
+logo192.png,1743977576018,883b9820cf90019f2d84dc41c5a9045a50c947110026fa9907e9b6de4b468143
+apple-touch-icon.png,1743977576003,6d6a75bece03595f34c90f74b45cf086588496aa8aeeae610ad55f294959fc8e
+algospectra-logo - Copy.jpeg,1743977576003,883b9820cf90019f2d84dc41c5a9045a50c947110026fa9907e9b6de4b468143
+static/media/god-dna.a807991fe1af48598719.jfif,1743980848896,0d07a05b971b8521f465e162853557910d3930e94ffe15ac7f7926a7b349750c
+android-chrome-512x512.png,1743977576003,9cc4ddd3ad0e6472d2d5548ef3fd4809fc5600dbbefa129d91f3b864ed4c1619
+static/js/main.25fbd85a.js,1743980848896,71f7f4fffaae694b270f0a8b6d4e9e1fd863226942071ed6c82093c78a4e7fb2
+static/js/main.25fbd85a.js.map,1743980848912,3b19ca0115bdacd7419cfa74ac7b90c775a1cd95f61b76a9781426c5f1a72e5c
