+++ src/styles/main.css
@@ -88,7 +88,8 @@ body {
   font-size: 0.7rem;
   font-weight: bold;
   padding: 0.6rem 1rem;
-  transform: rotate(-45deg);
+  border-radius: 40px;
+  /* transform: rotate(-45deg); */
   transform-origin: left top;
   box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
   z-index: 1;
